feat: adding docker in application
bug: video camera need link to browser
